+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 =========
 
 Detailed version changes.
+You can also view the changelog of the original library, fortnitepy, `here <https://fortnitepy.readthedocs.io/en/latest/changelog.html>`_.
 
 
 v0.0.4